Forms split and modularized.
